Update to Markdown Integration
This commit adds markdown to the timeline view and adds style so that
headers are all the same size. Also, this will change the timeline
so that you can click anywhere on the LI to go to the detailed view.
Finally this commit adds a preview button to the Moment form parital
so the user can preview their Markdown.